components/voice_agent_pipecat/pyproject.toml
[project]
name = "voice-agent-pipecat"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pipecat-ai[google,deepgram,local-smart-turn-v3,openai,silero,websocket]==0.0.101",
  "python-dotenv>=1.0.0",
  "loguru>=0.7.3",
]

[dependency-groups]
dev = [
  "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.uv.sources]
pipecat-ai = { git = "https://github.com/pipecat-ai/pipecat.git" }
